fixed paths in process-builders due to previous refactorings, also moved AnswerObject to other package
