+++ src/main/java/dp/resolver/parse/FactBuilder.java
@@ -28,7 +28,7 @@ public class FactBuilder {
     private List<CallNode> conflictNodes;
     private StringBuilder factsBuilder;
     private FileWriter writer;
-    private final static String ROOT_DIR = System.getProperty("user.dir") + "/src/main/java/dp/conflict/resolver/asp";
+    private final static String ROOT_DIR = System.getProperty("user.dir") + "/src/main/java/dp/resolver/asp";
     private File factsFile;
     private Map<String, Integer> idMap;
     private Set<String> alreadyLoadedJars;
